Population Module

The filter module provides an interface to select a data set, variable, and a query based on the value.

Format

R6 class

Input

datalist

A list of data sets. Can be tibbles, data frames, etc. Expects them to have a common ID field (default is USUBJID)

options

The options for the module of type reactiveValues with values

populationNumber

The index of the population that we are creating. If given, will include as the default name for the population "New population: populationNumber"

hideOnSave

Boolean. Should the population card be hidden when we save? Default: TRUE

idvar

ID field that is common across all datasets in datalist. Default USUBJID

Output

savePopulation

S3 object PopulationFilter returned when Save Population button is pressed

editPopulation

S3 object PopulationFilter which is updated after any change to the population.

See also

Super class

tidymodules::TidyModule -> Subpopulation

Methods

Public methods

Subpopulation$new()

Subpopulation$ui()

Subpopulation$server()

Subpopulation$clone()

Inherited methods

tidymodules::TidyModule$addInputPort()

tidymodules::TidyModule$addOutputPort()

tidymodules::TidyModule$assignPort()

tidymodules::TidyModule$callModule()

tidymodules::TidyModule$countInputPort()

tidymodules::TidyModule$countOutputPort()

tidymodules::TidyModule$deepClone()

tidymodules::TidyModule$definePort()

tidymodules::TidyModule$doServer()

tidymodules::TidyModule$execInput()

tidymodules::TidyModule$execOutput()

tidymodules::TidyModule$getGlobalSession()

tidymodules::TidyModule$getInput()

tidymodules::TidyModule$getInputPort()

tidymodules::TidyModule$getInputPorts()

tidymodules::TidyModule$getOutput()

tidymodules::TidyModule$getOutputPort()

tidymodules::TidyModule$getOutputPorts()

tidymodules::TidyModule$getPortDef()

tidymodules::TidyModule$getPortName()

tidymodules::TidyModule$getSession()

tidymodules::TidyModule$getSessionId()

tidymodules::TidyModule$getShinyInput()

tidymodules::TidyModule$getShinyOutput()

tidymodules::TidyModule$getShinySession()

tidymodules::TidyModule$getStore()

tidymodules::TidyModule$iport()

tidymodules::TidyModule$isGlobal()

tidymodules::TidyModule$isStored()

tidymodules::TidyModule$ns()

tidymodules::TidyModule$oport()

tidymodules::TidyModule$print()

tidymodules::TidyModule$render()

tidymodules::TidyModule$reset()

tidymodules::TidyModule$updateInputPort()

tidymodules::TidyModule$updateInputPorts()

tidymodules::TidyModule$updateOutputPort()

tidymodules::TidyModule$updateOutputPorts()

tidymodules::TidyModule$use()

Method new()

Usage

Subpopulation$new(...)

Method ui()

Usage

Subpopulation$ui()

Method server()

Usage

Subpopulation$server(input, output, session, data = NULL, options = NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

Subpopulation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
library(tidymodules)
library(subpat)
library(bs4Dash)

# Create a new population edit module

ui <- tagList(
  shinyjs::useShinyjs(),
  bs4DashPage(
    sidebar = bs4DashSidebar(disable = TRUE),
    body = bs4DashBody(
      popedit$ui(),

      fluidRow(
        column(width = 6,
               p("Editing population"),
               verbatimTextOutput('editing_population')),
        column(width = 6,
               p("Saved object"),
               verbatimTextOutput('saved_population'))
      )
    )
  )
)

# Define server logic required to draw a histogram
server <- function(input, output) {
  datalist <- reactive({
    list(
      "ADSL" = data.frame(
        a = c("x", "z", "y", "x", "z", "x", "x", "y", "x", "y"),
        b = c(29L, 4L, 15L, 28L, 41L, 25L, 41L, 15L, 6L, 17L),
        USUBJID = as.character(seq_len(10)),
        stringsAsFactors = FALSE
      ),
      "ADAE" = data.frame(
        a = c("x", "z", "y", "x", "z", "x", "x", "y", "x", "y"),
        c = c(1, NA, NA, 1, 2, NA, NA, 4, 3, 1),
        dates = seq(as.Date("2019-07-10"), as.Date("2019-08-12"), length.out = 10),
        USUBJID = as.character(seq_len(10)),
        stringsAsFactors = FALSE
      )
    )
  })

  popedit$callModule()

  opts <- reactiveValues(
    hideOnSave = FALSE,
    populationNumber = 1,
    subjectDs = "ADSL",
    idvar = "USUBJID"
  )

  observe({
    # Add the data to the module
    datalist %>1% popedit

    # Add the options to the population editing module
    opts %>2% popedit
  })

  output$saved_population <- renderPrint({
    req(popedit)
    str(popedit$getOutput(1)())
  })

  output$editing_population <- renderPrint({
    req(popedit)
    str(popedit$getOutput(2)())
  })
}

# Run the application 
shinyApp(ui = ui, server = server)
}