Time to event (TTE) module

Shiny TidyModule to perform exploratory survival analysis. Features Kaplan Meier survival function estimation and event summaries as well as Cox Proportional Hazard models.

Format

R6 class

See also

Super class

tidymodules::TidyModule -> TTE

Methods

Public methods

TTE$new()

TTE$coxUi()

TTE$coxForestUi()

TTE$coxVariablesUi()

TTE$plotUi()

TTE$eventTableUi()

TTE$standardUi()

TTE$ui()

TTE$server()

TTE$clone()

Inherited methods

tidymodules::TidyModule$addInputPort()

tidymodules::TidyModule$addOutputPort()

tidymodules::TidyModule$assignPort()

tidymodules::TidyModule$callModule()

tidymodules::TidyModule$countInputPort()

tidymodules::TidyModule$countOutputPort()

tidymodules::TidyModule$deepClone()

tidymodules::TidyModule$definePort()

tidymodules::TidyModule$doServer()

tidymodules::TidyModule$execInput()

tidymodules::TidyModule$execOutput()

tidymodules::TidyModule$getGlobalSession()

tidymodules::TidyModule$getInput()

tidymodules::TidyModule$getInputPort()

tidymodules::TidyModule$getInputPorts()

tidymodules::TidyModule$getOutput()

tidymodules::TidyModule$getOutputPort()

tidymodules::TidyModule$getOutputPorts()

tidymodules::TidyModule$getPortDef()

tidymodules::TidyModule$getPortName()

tidymodules::TidyModule$getSession()

tidymodules::TidyModule$getSessionId()

tidymodules::TidyModule$getShinyInput()

tidymodules::TidyModule$getShinyOutput()

tidymodules::TidyModule$getShinySession()

tidymodules::TidyModule$getStore()

tidymodules::TidyModule$iport()

tidymodules::TidyModule$isGlobal()

tidymodules::TidyModule$isStored()

tidymodules::TidyModule$ns()

tidymodules::TidyModule$oport()

tidymodules::TidyModule$print()

tidymodules::TidyModule$render()

tidymodules::TidyModule$reset()

tidymodules::TidyModule$updateInputPort()

tidymodules::TidyModule$updateInputPorts()

tidymodules::TidyModule$updateOutputPort()

tidymodules::TidyModule$updateOutputPorts()

tidymodules::TidyModule$use()

Method new()

Usage

TTE$new(...)

Method coxUi()

Usage

TTE$coxUi()

Method coxForestUi()

Usage

TTE$coxForestUi()

Method coxVariablesUi()

Usage

TTE$coxVariablesUi()

Method plotUi()

Usage

TTE$plotUi()

Method eventTableUi()

Usage

TTE$eventTableUi()

Method standardUi()

Usage

TTE$standardUi()

Method ui()

Usage

TTE$ui()

Method server()

Usage

TTE$server(input, output, session)

Method clone()

The objects of this class are cloneable with this method.

Usage

TTE$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
library(shiny)
library(subpat)
library(tidymodules)

tteMappingModule <- TTEMapping$new()
tteModule <- TTE$new()

ui <- fluidPage(
  titlePanel("TTE Analysis"),

  sidebarLayout(
    sidebarPanel(
      selectInput('dataset', 'survival dataset', choices = data(package = "survival")$results[, "Item"], selected = "lung"),
      tteMappingModule$ui()
    ),
    mainPanel(
      # Use the base shiny UI
      tteModule$standardUi()
    )
  )
)

server <- function(input, output, session) {

  options <- reactiveValues(
    makePlotly = FALSE,
    conftype = "log-log"
  )

  optionsMapping <- reactiveValues(
    population = FALSE,
    parameter = FALSE,
    parameter_value = FALSE,
    adam = FALSE
  )

  tteMappingModule$callModule()
  tteModule$callModule()

  # Load the data set from the survival package
  data_reactive <- reactive({
    req(input$dataset)
    ds <- trimws(gsub("\\(.*\\)", "", input$dataset))
    data(list = ds, package = "survival")

    # Reset the modules
    tteMappingModule <- TTEMapping$new()
    tteModule <- TTE$new()
    tteMappingModule$callModule()
    tteModule$callModule()

    get(ds)
  })

  observe({
    options %>4% tteModule
    optionsMapping %>2% tteMappingModule
    data_reactive %>1% tteModule
    # Get the mapping and pass into the TTE module
    data_reactive %>1% tteMappingModule %1>2% tteModule
  })
}

# Run the application
shinyApp(ui = ui, server = server)
}