Creates a event table from a survival model. Return a table with cumulative number of events, the number at risk, and the survival probability with confidence interval.

Creates a event table from a survival model. Return a table with cumulative number of events, the number at risk, and the survival probability with confidence interval.

kmEventTables(
  fit,
  times = round(seq(min(fit$time), ceiling(max(fit$time)), length.out = 10))
)

Arguments

fit
output from survival::survfit

times
Optional vector of times to show in the table. By default will include equally spaced times from the minimum event time to the maximum event time with 10 time points.

Value

kable output of the KM event table from survfit fit at time points times

Examples

library(survival)
fit <- survfit(Surv(time, status) ~ sex, data=survival::lung)
subpat::kmEventTables(fit)

#> <table class='gmisc_table' style='border-collapse: collapse; margin-top: 1em; margin-bottom: 1em; width: 850px; ' >
#> <thead>
#> <tr>
#> <th colspan='1' style='font-weight: 900; border-bottom: 1px solid grey; border-top: 2px solid grey; text-align: center;'> </th><th style='border-top: 2px solid grey;; border-bottom: hidden;'>&nbsp;</th>
#> <th colspan='3' style='font-weight: 900; border-bottom: 1px solid grey; border-top: 2px solid grey; text-align: center;'>sex=1</th><th style='border-top: 2px solid grey;; border-bottom: hidden;'>&nbsp;</th>
#> <th colspan='3' style='font-weight: 900; border-bottom: 1px solid grey; border-top: 2px solid grey; text-align: center;'>sex=2</th>
#> </tr>
#> <tr>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>Time</th>
#> <th style='border-bottom: 1px solid grey;' colspan='1'>&nbsp;</th>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>Cum. Events</th>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>At Risk</th>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>Survival Prob. [95% CI]</th>
#> <th style='border-bottom: 1px solid grey;' colspan='1'>&nbsp;</th>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>Cum. Events</th>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>At Risk</th>
#> <th style='border-bottom: 1px solid grey; text-align: center;'>Survival Prob. [95% CI]</th>
#> </tr>
#> </thead>
#> <tbody>
#> <tr>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>5</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>0</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>138</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>100.0 [100.0 100.0]</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>1</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>90</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>98.9 [96.7 100.0]</td>
#> </tr>
#> <tr style='background-color: #f7f7f7;'>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>118</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>29</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>110</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>79.0 [72.5 86.1]</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>8</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>80</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>91.1 [85.4 97.2]</td>
#> </tr>
#> <tr>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>231</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>63</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>65</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>53.6 [45.8 62.7]</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>22</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>58</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>74.5 [65.8 84.3]</td>
#> </tr>
#> <tr style='background-color: #f7f7f7;'>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>344</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>81</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>39</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>37.4 [29.8 47.1]</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>30</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>38</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>62.6 [52.7 74.5]</td>
#> </tr>
#> <tr>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>457</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>94</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>24</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>24.6 [17.7 34.1]</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>41</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>22</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>43.1 [32.4 57.2]</td>
#> </tr>
#> <tr style='background-color: #f7f7f7;'>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>570</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>101</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>15</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>16.7 [10.8 25.8]</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>45</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>12</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>34.3 [23.9 49.3]</td>
#> </tr>
#> <tr>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>683</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>107</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>9</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>10.0 [5.5 18.3]</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>47</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>9</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>28.1 [17.8 44.3]</td>
#> </tr>
#> <tr style='background-color: #f7f7f7;'>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>796</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>110</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>6</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>6.7 [3.1 14.3]</td>
#> <td style='background-color: #f7f7f7;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>53</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>2</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; text-align: center;'>8.3 [2.6 27.0]</td>
#> </tr>
#> <tr>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>909</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>112</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>2</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>3.6 [1.1 11.7]</td>
#> <td style='' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>53</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>1</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; text-align: center;'>8.3 [2.6 27.0]</td>
#> </tr>
#> <tr style='background-color: #f7f7f7;'>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'>1022</td>
#> <td style='background-color: #f7f7f7; border-bottom: 2px solid grey;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'>112</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'>1</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'>3.6 [1.1 11.7]</td>
#> <td style='background-color: #f7f7f7; border-bottom: 2px solid grey;' colspan='1'>&nbsp;</td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'></td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'></td>
#> <td style='padding-left: 1.5em; padding-right: 1.5em; background-color: #f7f7f7; border-bottom: 2px solid grey; text-align: center;'></td>
#> </tr>
#> </tbody>
#> </table>